+++ src/errors/NotFound.ts
@@ -13,6 +13,6 @@ export class NotFound extends CustomError {
     this.errorType = errorType || `Application.${this.name}`;
 
     // Additional error detail
-    this.detail = 'The server has not found anything matching the Request-URI. No indication is given of whether the condition is temporary or permanent.';
+    this.detail = detail || 'The requested resource could not be found but may be available again in the future.';
   }
 }