+++ src/index.ts
@@ -1,5 +1,7 @@
 export { BadGateway } from './errors/BadGateway';
 export { BadRequest } from './errors/BadRequest';
 export { CustomError } from './CustomError';
+export { Forbidden } from './Forbidden';
 export { NotFound } from './errors/NotFound';
 export { RecordNotFound } from './errors/RecordNotFound';
+export { Unauthorized } from './errors/Unauthorized';